move util function into it's own file
